refactor(edge): move url and tunnel fields
