+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/shared/EdgeAgentTab/EdgeAgentForm/EdgeAgentFieldset.tsx
@@ -1,9 +1,8 @@
 import { isBE } from '@/react/portainer/feature-flags/feature-flags.service';
+import { PortainerTunnelAddrField } from '@/react/portainer/common/PortainerTunnelAddrField';
+import { PortainerUrlField } from '@/react/portainer/common/PortainerUrlField';
 
 import { NameField } from '../../NameField';
-
-import { PortainerTunnelAddrField } from './PortainerTunnelAddrField';
-import { PortainerUrlField } from './PortainerUrlField';
 
 interface EdgeAgentFormProps {
   readonly?: boolean;
